Fixes #101 - Use temporary build_dir and pass --output directly in youtube2zim-playlists

+++ youtube2zim/scraper.py
@@ -12,6 +12,7 @@ import os
 import json
 import locale
 import shutil
+import tempfile
 import subprocess
 import datetime
 import functools
@@ -69,6 +70,7 @@ class Youtube2Zim(object):
         no_zim,
         fname,
         debug,
+        build_dir,
         keep_build_dir,
         skip_download,
         max_concurrency,
@@ -117,8 +119,14 @@ class Youtube2Zim(object):
         self.main_color = main_color
         self.secondary_color = secondary_color
 
-        # process-related
+        # directory setup
         self.output_dir = Path(output_dir).expanduser().resolve()
+        self.temp_dir = tempfile.TemporaryDirectory() if not build_dir else None
+        self.custom_build_dir = (
+            Path(build_dir).expanduser().resolve() if build_dir else None
+        )
+
+        # process-related
         self.playlists = []
         self.uploads_playlist_id = None
         self.videos_ids = []
@@ -131,8 +139,6 @@ class Youtube2Zim(object):
         self.keep_build_dir = keep_build_dir
         self.skip_download = skip_download
         self.max_concurrency = max_concurrency
-
-        self.build_dir = self.output_dir.joinpath("build")
 
         # store ZIM-related info
         self.zim_info = ZimInfo(
@@ -179,6 +185,12 @@ class Youtube2Zim(object):
     @property
     def assets_src_dir(self):
         return self.templates_dir.joinpath("assets")
+
+    @property
+    def build_dir(self):
+        return (
+            self.custom_build_dir if self.custom_build_dir else Path(self.temp_dir.name)
+        )
 
     @property
     def assets_dir(self):
@@ -256,9 +268,6 @@ class Youtube2Zim(object):
             f"starting youtube scraper for {self.collection_type}#{self.youtube_id}"
         )
         logger.info("preparing build folder at {}".format(self.build_dir.resolve()))
-        if not self.keep_build_dir and self.build_dir.exists():
-            shutil.rmtree(self.cache_dir, ignore_errors=True)
-            shutil.rmtree(self.build_dir)
         self.prepare_build_folder()
 
         logger.info("testing Youtube credentials")
@@ -335,6 +344,7 @@ class Youtube2Zim(object):
         self.make_html_files(succeeded)
 
         # make zim file
+        os.makedirs(self.output_dir, exist_ok=True)
         if not self.no_zim:
             period = datetime.datetime.now().strftime("%Y-%m")
             self.fname = (
@@ -345,11 +355,20 @@ class Youtube2Zim(object):
             logger.info("building ZIM file")
             print(self.zim_info.to_zimwriterfs_args())
             make_zim_file(self.build_dir, self.output_dir, self.fname, self.zim_info)
-            logger.info("removing HTML folder")
-            if not self.keep_build_dir:
-                shutil.rmtree(self.build_dir, ignore_errors=True)
-
+        self.post_process_build_dir()
         logger.info("all done!")
+
+    def post_process_build_dir(self):
+        """ Deletes/keeps build_dir according to user's choice """
+
+        if self.keep_build_dir and not self.custom_build_dir:
+            logger.info("Saving build directory")
+            shutil.copytree(
+                self.build_dir, self.output_dir.joinpath(f"{self.fname}_build")
+            )
+        elif not self.keep_build_dir and self.custom_build_dir:
+            logger.info("Removing build directory")
+            shutil.rmtree(self.build_dir, ignore_errors=True)
 
     def s3_credentials_ok(self):
         logger.info("testing S3 Optimization Cache credentials")
@@ -378,8 +397,13 @@ class Youtube2Zim(object):
     def prepare_build_folder(self):
         """ prepare build folder before we start downloading data """
 
-        # create build folder
-        os.makedirs(self.build_dir, exist_ok=True)
+        if self.custom_build_dir:
+            if not self.keep_build_dir and self.build_dir.exists():
+                shutil.rmtree(self.cache_dir, ignore_errors=True)
+                shutil.rmtree(self.build_dir)
+
+            # create build folder
+            os.makedirs(self.build_dir, exist_ok=True)
 
         # copy assets
         if self.assets_dir.exists():
